feat: Update Confluence Page Settings Command
This feature will enable users to update the settings that are used by the integration without needing to remember the settings. This also enables a single list of all configuration we support in the YAML frontmatter and how it works with validation and processing functions.

feat: Write `connie-publish: true` to all files that have been published to ensure even if you move the files they still will be published.