Remove first heading from page when using as page title
This commit modifies the processing of the page title in the `conniePerPageConfig` object. If the `firstHeadingPageTitle` setting is enabled and the first element of the page content is a heading, the text of this heading is now used as the page title. Additionally, to prevent duplication, the first heading is removed from the body of the page. This change only applies if the first heading's text content is not an empty string.